Prune logs when creating RollingFileAppender
For short-lived apps (think something like an aws lambda, or a cron job), the
number of log files can grow without bound, because the app is basically never
alive and logging when a rollover boundary hits. So the appender needs to ensure
we prune at _some_ point.

Startup is the cleanest and simplest place to do that; I tried some other
options (explicitly do it on first write, _implicitly_ do it on first write by
deliberately writing a wrong `next_date`...) but they all had more obnoxious
trade-offs.