TEMP: knock out path dep for tracing-subscriber
This makes it impossible to swap out the appender dependency in my own project
without patching all of its in-repo dependencies as well, which I don't want to
do. It turns out the crates.io service itself munges Cargo.toml to remove these
during ingress, which is why this isn't a problem in the published versions.